+++ backend/bottari/src/main/java/com/bottari/apiworkshop/dto/Cursor.java
@@ -1,4 +1,4 @@
-package com.bottari.apiworkshop;
+package com.bottari.apiworkshop.dto;
 
 public record Cursor(
         Long limit,
