+++ backend/bottari/src/main/java/com/bottari/apiworkshop/repository/OpenApiItemRepository.java
@@ -1,4 +1,4 @@
-package com.bottari.apiworkshop;
+package com.bottari.apiworkshop.repository;
 
 import com.bottari.bottaritemplate.domain.BottariTemplateItem;
 import java.util.List;
