+++ backend/bottari/src/main/java/com/bottari/apiworkshop/service/OpenApiService.java
@@ -1,5 +1,13 @@
-package com.bottari.apiworkshop;
+package com.bottari.apiworkshop.service;
 
+import com.bottari.apiworkshop.repository.OpenApiItemRepository;
+import com.bottari.apiworkshop.repository.OpenApiTemplateRepository;
+import com.bottari.apiworkshop.dto.Cursor;
+import com.bottari.apiworkshop.dto.CursorRequest;
+import com.bottari.apiworkshop.repository.ItemProjection;
+import com.bottari.apiworkshop.dto.ItemResponse;
+import com.bottari.apiworkshop.dto.MostIncludedQuery;
+import com.bottari.apiworkshop.dto.MostIncludedResponse;
 import com.bottari.bottaritemplate.domain.BottariTemplate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
