+++ backend/bottari/src/test/java/com/bottari/apiworkshop/OpenApiServiceTest.java
@@ -3,6 +3,11 @@ package com.bottari.apiworkshop;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import com.bottari.apiworkshop.dto.CursorRequest;
+import com.bottari.apiworkshop.dto.ItemResponse;
+import com.bottari.apiworkshop.dto.MostIncludedQuery;
+import com.bottari.apiworkshop.dto.MostIncludedResponse;
+import com.bottari.apiworkshop.service.OpenApiService;
 import com.bottari.bottaritemplate.domain.BottariTemplate;
 import com.bottari.bottaritemplate.domain.BottariTemplateItem;
 import com.bottari.config.JpaAuditingConfig;
